Gotta save them 2 bytes

+++ src/event.h
@@ -65,7 +65,7 @@ namespace t4editor {
             event_type type;
             vec2 cursorPosition;
             double scrollDelta;
-            unsigned int key;
+            unsigned short key;
     };
     
     class load_level_event : public event {
